refactor(#620): define settings schema in module
`__init__.py` was getting crowded
