+++ src/pyspry.pyi
@@ -6,9 +6,9 @@ from __future__ import annotations
 
 from pathlib import Path
 
-from config_ninja.settings import ConfigNinjaObject
-from config_ninja.settings import Dest as Dest
-from config_ninja.settings import Source as Source
+from config_ninja.settings.schema import ConfigNinjaObject
+from config_ninja.settings.schema import Dest as Dest
+from config_ninja.settings.schema import Source as Source
 
 # pylint: disable=unused-argument,missing-class-docstring,missing-function-docstring
